MAJ backend avec toutes les modifs

+++ backend/controllers/sauce.js
@@ -1,4 +1,5 @@
 const Sauce = require("../models/Sauce");
+const fs = require("fs");
 
 /**Renvoie un tableau de toutes les sauces */
 exports.getAllSauces = (req, res, next) => {
@@ -58,27 +59,41 @@ exports.likeSauce = (req, res, next) => {
 
 /**Met à jour la sauce */
 exports.updateSauce = (req, res, next) => {
-    const sauce = new Sauce({
-        userId: req.body._id,
-        name: req.body.name,
-        manufacturer: req.body.manufacturer,
-        description: req.body.description,
-        mainPepper: req.body.mainPepper,
-        imageUrl: req.body.imageUrl,
-        heat: req.body.heat,
-        likes: 0,
-        dislikes: 0,
-        usersLiked: [],
-        usersDisliked: [],
-    });
-    Sauce.updateOne({ _id: req.params.id }, sauce)
-        .then(() => { res.status(200).json({ message: "Votre sauce a bien été mise à jour !" }) })
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete sauceObject._userId;
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: "Pas autorisé" });
+            } else {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: "Votre sauce a bien été mise à jour !" }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
         .catch((error) => res.status(400).json({ error }));
 };
 
 /**Supprime la sauce */
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => { res.status(200).json({ message: "Votre sauce a bien été supprimée !" }) })
-        .catch((error) => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: "Pas autorisé" });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: "Votre sauce a bien été supprimée !" }) })
+                        .catch((error) => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
